added parameter to $cms->content()
Added a parameter, that when used, pre-populates the database with
content, this way you can easily add content to the page without going
to the browser and copy and pasting the information.

It also lets me create templates, that are easy to get up and running.
This might make adopting the CMS easier.